requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.8
django-allauth==0.58.2
django-countries==7.5.1
django-credit-cards==0.4.1
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-phone-field==1.8.1
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
